pyproject.toml
[tool.poetry]
name = "enelvo-microservice"
version = "0.10.0"
description = "Enelvo pt_BR normalizer microservice."
authors = ["Thiago Carvalho D'Ávila <thiagocavila@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/staticdev/enelvo-microservice"
repository = "https://github.com/staticdev/enelvo-microservice"

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.29.0"
protobuf = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
